Add a "make tests" command to the Makefiles

+++ wsf-scraper-web/tests/test_db_tools.py
@@ -1,5 +1,6 @@
 import unittest
 import os
+import logging
 from tools import DatabaseConnector
 
 
@@ -7,9 +8,17 @@ class TestDBTools(unittest.TestCase):
 
     def setUp(self):
         """Assuming the database is already set up."""
-        self.database = DatabaseConnector(
-            os.environ['DATABASE_URL_TEST']
-        )
+        database_url = os.getenv('DATABASE_URL_TEST')
+        self.logger = logging.getLogger(__name__)
+        self.logger.setLevel(logging.INFO)
+
+        if not database_url:
+            self.logger.warning('No DATABASE_URL_TEST in environment.'
+                                'Unable to unittest the database methods.')
+            self.database = None
+            return
+
+        self.database = DatabaseConnector(database_url)
         mock_publication = {
             'title': 'foo',
             'uri': 'http://foo.bar',
@@ -38,34 +47,37 @@ class TestDBTools(unittest.TestCase):
         self.database.insert_joints_and_text('keyword', keywords, id_pub)
 
     def tearDown(self):
-        self.database._execute('DELETE FROM publications_types')
-        self.database._execute('DELETE FROM publications_subjects')
-        self.database._execute('DELETE FROM publications_sections')
-        self.database._execute('DELETE FROM publications_keywords')
-        self.database._execute('DELETE FROM type')
-        self.database._execute('DELETE FROM subject')
-        self.database._execute('DELETE FROM section')
-        self.database._execute('DELETE FROM keyword')
-        self.database._execute('DELETE FROM publication')
-        self.database._execute('DELETE FROM provider')
+        if self.database:
+            self.database._execute('DELETE FROM publications_types')
+            self.database._execute('DELETE FROM publications_subjects')
+            self.database._execute('DELETE FROM publications_sections')
+            self.database._execute('DELETE FROM publications_keywords')
+            self.database._execute('DELETE FROM type')
+            self.database._execute('DELETE FROM subject')
+            self.database._execute('DELETE FROM section')
+            self.database._execute('DELETE FROM keyword')
+            self.database._execute('DELETE FROM publication')
+            self.database._execute('DELETE FROM provider')
 
     def test_full_publication(self):
-        self.assertTrue(self.database.get_scraping_info('0' * 32))
+        if self.database:
+            self.assertTrue(self.database.get_scraping_info('0' * 32))
 
     def test_joints(self):
-        self.database._execute('SELECT * FROM section')
-        self.assertTrue(self.database.cursor.fetchone())
-        self.database._execute('SELECT * FROM keyword')
-        self.assertTrue(self.database.cursor.fetchone())
-        self.database._execute('SELECT * FROM publications_sections')
-        self.assertTrue(self.database.cursor.fetchone())
-        self.database._execute('SELECT * FROM publications_keywords')
-        self.assertTrue(self.database.cursor.fetchone())
-        self.database._execute('SELECT * FROM type')
-        self.assertTrue(self.database.cursor.fetchone())
-        self.database._execute('SELECT * FROM publications_types')
-        self.assertTrue(self.database.cursor.fetchone())
-        self.database._execute('SELECT * FROM subject')
-        self.assertTrue(self.database.cursor.fetchone())
-        self.database._execute('SELECT * FROM publications_subjects')
-        self.assertTrue(self.database.cursor.fetchone())
+        if self.database:
+            self.database._execute('SELECT * FROM section')
+            self.assertTrue(self.database.cursor.fetchone())
+            self.database._execute('SELECT * FROM keyword')
+            self.assertTrue(self.database.cursor.fetchone())
+            self.database._execute('SELECT * FROM publications_sections')
+            self.assertTrue(self.database.cursor.fetchone())
+            self.database._execute('SELECT * FROM publications_keywords')
+            self.assertTrue(self.database.cursor.fetchone())
+            self.database._execute('SELECT * FROM type')
+            self.assertTrue(self.database.cursor.fetchone())
+            self.database._execute('SELECT * FROM publications_types')
+            self.assertTrue(self.database.cursor.fetchone())
+            self.database._execute('SELECT * FROM subject')
+            self.assertTrue(self.database.cursor.fetchone())
+            self.database._execute('SELECT * FROM publications_subjects')
+            self.assertTrue(self.database.cursor.fetchone())
